refactor(outbound): inline delivery diagnostics
