+++ src/infra/outbound/deliver-payload.ts
@@ -12,11 +12,6 @@ import {
 import { createSubsystemLogger } from "../../logging/subsystem.js";
 import type { OutboundMediaAccess } from "../../media/load-options.js";
 import { resolveAgentScopedOutboundMediaAccess } from "../../media/read-capability.js";
-import { diagnosticErrorCategory } from "../diagnostic-error-metadata.js";
-import {
-  emitInternalDiagnosticEvent as emitDiagnosticEvent,
-  type DiagnosticMessageDeliveryKind,
-} from "../diagnostic-events.js";
 import { formatErrorMessage } from "../errors.js";
 import type {
   ChannelHandler,
@@ -25,23 +20,14 @@ import type {
 } from "./deliver-contracts.js";
 import type { OutboundDeliveryResult, OutboundPayloadDeliveryKind } from "./deliver-types.js";
 import { flattenMarkdownDetails } from "./markdown-details.js";
-import type { DeliveryMirror } from "./mirror.js";
 import {
   summarizeOutboundPayloadForTransport,
   type NormalizedOutboundPayload,
   type OutboundPayloadPlan,
 } from "./payloads.js";
 import { stripInternalRuntimeScaffolding } from "./protocol-scaffolding.js";
-import type { OutboundSessionContext } from "./session-context.js";
 
 const log = createSubsystemLogger("outbound/deliver");
-
-export function sessionKeyForDeliveryDiagnostics(params: {
-  mirror?: DeliveryMirror;
-  session?: OutboundSessionContext;
-}): string | undefined {
-  return params.mirror?.sessionKey ?? params.session?.key ?? params.session?.policyKey;
-}
 
 export function deliveryKindForPayload(
   payload: ReplyPayload,
@@ -54,53 +40,6 @@ export function deliveryKindForPayload(
     return "other";
   }
   return "text";
-}
-
-export function emitMessageDeliveryStarted(params: {
-  channel: string;
-  deliveryKind: DiagnosticMessageDeliveryKind;
-  sessionKey?: string;
-}): void {
-  emitDiagnosticEvent({
-    type: "message.delivery.started",
-    channel: params.channel,
-    deliveryKind: params.deliveryKind,
-    ...(params.sessionKey ? { sessionKey: params.sessionKey } : {}),
-  });
-}
-
-export function emitMessageDeliveryCompleted(params: {
-  channel: string;
-  deliveryKind: DiagnosticMessageDeliveryKind;
-  durationMs: number;
-  resultCount: number;
-  sessionKey?: string;
-}): void {
-  emitDiagnosticEvent({
-    type: "message.delivery.completed",
-    channel: params.channel,
-    deliveryKind: params.deliveryKind,
-    durationMs: params.durationMs,
-    resultCount: params.resultCount,
-    ...(params.sessionKey ? { sessionKey: params.sessionKey } : {}),
-  });
-}
-
-export function emitMessageDeliveryError(params: {
-  channel: string;
-  deliveryKind: DiagnosticMessageDeliveryKind;
-  durationMs: number;
-  error: unknown;
-  sessionKey?: string;
-}): void {
-  emitDiagnosticEvent({
-    type: "message.delivery.error",
-    channel: params.channel,
-    deliveryKind: params.deliveryKind,
-    durationMs: params.durationMs,
-    errorCategory: diagnosticErrorCategory(params.error),
-    ...(params.sessionKey ? { sessionKey: params.sessionKey } : {}),
-  });
 }
 
 export function normalizeEmptyPayloadForDelivery(payload: ReplyPayload): ReplyPayload | null {
